style: move to utils for duration to timestamp format function
