--- packages/script/zenza-watch/src/util.ts
@@ -1,11 +1,0 @@
-let commonHeader: HTMLDivElement | null;
-
-export function computeBaseZIndex() {
-  commonHeader ||= document.querySelector("div#CommonHeader");
-
-  if (commonHeader == null) {
-    return "5000000";
-  }
-
-  return getComputedStyle(commonHeader).zIndex;
-}
